# Improve jar downloading performance