* Use direct IO.copy to perform a streaming copy.
* Don't perform a checksum on every download. Instead compute it once and use it for the rest of the downloads.