@ketan 2016-10-25
### Before

The CPU is pegged at about 85% when the test runs.

#### Warmup run (simulate a server upgrade)


```
$ ab -q -c200 -n500 http://localhost:8253/go/admin/agent
This is ApacheBench, Version 2.3 <$Revision: 1706008 $>
Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
Licensed to The Apache Software Foundation, http://www.apache.org/

Benchmarking localhost (be patient).....done


Server Software:
Server Hostname:        localhost
Server Port:            8253

Document Path:          /go/admin/agent
Document Length:        29235047 bytes

Concurrency Level:      200
Time taken for tests:   15.073 seconds
Complete requests:      500
Failed requests:        0
Total transferred:      14617654463 bytes
HTML transferred:       14617523500 bytes
Requests per second:    33.17 [#/sec] (mean)
Time per request:       6029.056 [ms] (mean)
Time per request:       30.145 [ms] (mean, across all concurrent requests)
Transfer rate:          947083.74 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    3   1.6      3       7
Processing:  2120 5502 1361.8   6155    6788
Waiting:        2  101 109.0     30     274
Total:       2121 5506 1362.5   6158    6794

Percentage of the requests served within a certain time (ms)
  50%   6158
  66%   6257
  75%   6322
  80%   6345
  90%   6393
  95%   6436
  98%   6506
  99%   6583
 100%   6794 (longest request)
```

#### Subsequent runs (this is a rare case, since it implies that all agents were rebooted at the same time)

Not much of an improvement here. Output went from 33 request per seconds to 36 requests per second.x

```
$ ab -q -c200 -n500 http://localhost:8253/go/admin/agent
This is ApacheBench, Version 2.3 <$Revision: 1706008 $>
Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
Licensed to The Apache Software Foundation, http://www.apache.org/

Benchmarking localhost (be patient).....done


Server Software:
Server Hostname:        localhost
Server Port:            8253

Document Path:          /go/admin/agent
Document Length:        29235047 bytes

Concurrency Level:      200
Time taken for tests:   13.655 seconds
Complete requests:      500
Failed requests:        0
Total transferred:      14617654491 bytes
HTML transferred:       14617523500 bytes
Requests per second:    36.62 [#/sec] (mean)
Time per request:       5461.991 [ms] (mean)
Time per request:       27.310 [ms] (mean, across all concurrent requests)
Transfer rate:          1045410.20 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    4   2.2      3      15
Processing:  2163 4986 1207.8   5606    5862
Waiting:        2   74  88.2      8     198
Total:       2165 4990 1208.2   5611    5865

Percentage of the requests served within a certain time (ms)
  50%   5611
  66%   5693
  75%   5727
  80%   5749
  90%   5801
  95%   5837
  98%   5847
  99%   5857
 100%   5865 (longest request)
```

### After

The CPU is pegged at about 85% when the test runs.

#### Warmup run (simulate a server upgrade)

```
$ ab -q -c200 -n500 http://localhost:8153/go/admin/agent
This is ApacheBench, Version 2.3 <$Revision: 1706008 $>
Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
Licensed to The Apache Software Foundation, http://www.apache.org/

Benchmarking localhost (be patient).....done


Server Software:
Server Hostname:        localhost
Server Port:            8153

Document Path:          /go/admin/agent
Document Length:        29237685 bytes

Concurrency Level:      200
Time taken for tests:   11.315 seconds
Complete requests:      500
Failed requests:        0
Total transferred:      14618997999 bytes
HTML transferred:       14618842500 bytes
Requests per second:    44.19 [#/sec] (mean)
Time per request:       4526.180 [ms] (mean)
Time per request:       22.631 [ms] (mean, across all concurrent requests)
Transfer rate:          1261670.13 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        1    4   2.4      4      15
Processing:  1597 4172 1114.2   4787    4911
Waiting:        2  125 223.6     11     789
Total:       1606 4177 1114.1   4793    4917

Percentage of the requests served within a certain time (ms)
  50%   4793
  66%   4877
  75%   4884
  80%   4886
  90%   4912
  95%   4914
  98%   4915
  99%   4915
 100%   4917 (longest request)
```

#### Subsequent runs (this is a rare case, since it implies that all agents were rebooted at the same time)

Not much of an improvement here. Output went from 44 request per seconds to 49 requests per second.

```
ab -q -c200 -n500 http://localhost:8153/go/admin/agent
This is ApacheBench, Version 2.3 <$Revision: 1706008 $>
Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
Licensed to The Apache Software Foundation, http://www.apache.org/

Benchmarking localhost (be patient).....done


Server Software:
Server Hostname:        localhost
Server Port:            8153

Document Path:          /go/admin/agent
Document Length:        29237685 bytes

Concurrency Level:      200
Time taken for tests:   10.050 seconds
Complete requests:      500
Failed requests:        0
Total transferred:      14618997988 bytes
HTML transferred:       14618842500 bytes
Requests per second:    49.75 [#/sec] (mean)
Time per request:       4020.199 [ms] (mean)
Time per request:       20.101 [ms] (mean, across all concurrent requests)
Transfer rate:          1420463.47 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        1    4   2.7      4      19
Processing:  1347 3604 843.5   3909    4304
Waiting:        2   25  32.3      6     102
Total:       1351 3608 843.7   3915    4308

Percentage of the requests served within a certain time (ms)
  50%   3915
  66%   4161
  75%   4208
  80%   4209
  90%   4216
  95%   4217
  98%   4217
  99%   4218
 100%   4308 (longest request)
```

## Next steps

Interestingly moving the agent.jar to the `assets` directory so as to have it handled by the `AssetsContextHandler` gets the performance to around 95 requests per minute. Additionally the CPU is pegged at about 50% instead of the 85% that was noticed earlier.