@arvindsv 2016-10-25
The idea looks fine to me. Why wouldn't we, then, move it to `assets/`?