@ketan 2016-10-25
The intention of moving under `assets` was to validate the performance of `org.eclipse.jetty.server.handler.ResourceHandler`. It seems to connect the channel from the file resource, directly to the channel for the http connection.

I'll submit another PR to move it under a different "handler" later, preferably with a "filter" to setup the relevant headers for checksums. Ideally I'd want to not have the filter and just use plain `Etag` or `Modified`

To improve this situation further, I'd probably want to use some form of a proper `ETag` or `Last-Modified` headers that http clients understand out of the box, instead of the hacky `Content-MD5`. This will allow http caching libraries to do the right thing without us having to re-invent the wheel using the `Content-MD5` header.