@arvindsv 2016-10-25
> To improve this situation further, I'd probably want to use some form of a proper ETag or Last-Modified headers that http clients understand out of the box, instead of the hacky Content-MD5. This will allow http caching libraries to do the right thing without us having to re-invent the wheel using the Content-MD5 header.

👍